refactor: remove reset password logic from UserContext and implement in ResetPasswordForm
